tests/ui/requirements.txt
PyPOM==1.0
pytest==2.9.2
pytest-selenium==1.3.1
